Added DFS algorithm
There is 4 new search time.
To discover all the Vertices there is DFS Iterative and Recursive.
To find all vertices connected to a source key there is DFS Iterative and Recursive.
Recursive is tested with smaller number as it generate a stack overflow if the amount of vertices and edges is too large.
All the same test are still performed.
tweaks to array populating function.